Move cli into root dir and lib into temple/temple.
The reason for this change is that I am planning to rewrite the generated code so that it does not require you to import the lib. That means that the cli will be more important, as the most common use case will be to use the cli and not import the lib at all. Since it is more important, it gets the root dir so installs are shorter (go get github.com/go-humble/temple instead of go get github.com/go-humble/temple/cmd/temple)